fctl/nav: Move NavBar into nav directory
